Pixelrift/requirements.txt
asyncpg==0.25.0
discord.py==2.0.0
